block user done update other cache implementation

+++ cache/posts/postDeleteCache.js
@@ -1,23 +1,22 @@
 // to delete post entry from cache
 
 const redisClient = require('../redis.js');
-const Post = require('../../models/post.js');
 
-const { getIndexKey } = require('../../utils/cache/cacheIndexUtils.js');
+const { getUserPostKey, getPostIdPath } = require('./postCache.js');
 const { getPostCommentKey } = require('../../cache/comments/commentCache.js');
 const { deleteAllCachedComments } = require('../../cache/comments/commentDeleteCache.js');
 
-// to delete post entry from cache and database
-function deleteCachedPost(key, postIndex, postId) {
-    const postIdKey = getIndexKey(key);
+// to delete post entry from cache
+function deleteCachedPost(userId, postId) {
+    if (!redisClient.isReady) {
+        return [];
+    }
 
-    // remove post entry and associated comments from cache and database
-    return Promise.all([
-        redisClient.json.arrPop(key, '$', postIndex),
-        redisClient.json.arrPop(postIdKey, '$', postIndex),
-        Post.findByIdAndDelete(postId),
+    // remove post entry and associated comments from cache
+    return [
+        redisClient.json.del(getUserPostKey(userId), getPostIdPath(postId)),
         deleteAllCachedComments(getPostCommentKey(postId))
-    ]);
+    ];
 }
 
 module.exports = {
